Feature/melhorias geral

+++ app/Controllers/Http/CategoryController.js
@@ -1,74 +1,26 @@
-'use strict';
-
-/** @typedef {import('@adonisjs/framework/src/Request')} Request */
-/** @typedef {import('@adonisjs/framework/src/Response')} Response */
-/** @typedef {import('@adonisjs/framework/src/View')} View */
+/** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
+const Employee = use('App/Models/Employee');
+/** @type {typeof import('@adonisjs/lucid/src/Lucid/Model')} */
 const Category = use('App/Models/Category');
-/**
- * Resourceful controller for interacting with categories
- */
+
 class CategoryController {
-  /**
-   * Show a list of all categories.
-   * GET categories
-   *
-   * @param {object} ctx
-   * @param {Request} ctx.request
-   * @param {Response} ctx.response
-   * @param {View} ctx.view
-   */
   async index({ response }) {
     const categories = await Category.query().fetch();
     response.status(200).json(categories);
   }
 
-  /**
-   * Create/save a new category.
-   * POST categories
-   *
-   * @param {object} ctx
-   * @param {Request} ctx.request
-   * @param {Response} ctx.response
-   */
-  // "label": "Cereais",
-  // "parent_id": null,
-  // "child_count": 2,
-  // "created_at": "16/06/2020 18:24:00",
-  // "updated_at": "16/06/2020 18:24:00"
-  async store({ request, response }) {
-    const data = request.only([
-      'level',
-      'label',
-      'parent_id',
-      'child_count'
-    ]);
+  async store({ request }) {
+    const data = request.only(['level', 'label', 'parent_id', 'child_count']);
 
     const category = await Category.create(data);
     return category;
   }
 
-  /**
-   * Display a single category.
-   * GET categories/:id
-   *
-   * @param {object} ctx
-   * @param {Request} ctx.request
-   * @param {Response} ctx.response
-   * @param {View} ctx.view
-   */
-  async show({ params, request, response, view }) {
-    let category = await Category.findOrFail(params.id);
+  async show({ params }) {
+    const category = await Category.findOrFail(params.id);
     return category;
   }
 
-  /**
-   * Update category details.
-   * PUT or PATCH categories/:id
-   *
-   * @param {object} ctx
-   * @param {Request} ctx.request
-   * @param {Response} ctx.response
-   */
   async update({ params, request, response }) {
     const category = await Category.findOrFail(params.id);
 
@@ -79,25 +31,20 @@ class CategoryController {
     return response.status(200).json(category);
   }
 
-  /**
-   * Delete a category with id.
-   * DELETE categories/:id
-   *
-   * @param {object} ctx
-   * @param {Request} ctx.request
-   * @param {Response} ctx.response
-   */
-  async destroy({ params, request, response }) {
+  async destroy({ params, response, auth }) {
     const category = await Category.findOrFail(params.id);
 
-    // if (category.user_id !== auth.user.id) {
-    //   return response.status(401).send({ error: 'Not authorized' })
-    // }
+    const admExists = await Employee.findByOrFail('user_id', auth.user.id);
+
+    if (admExists.type !== 'ADM') {
+      return response.status(401).json({
+        unauthorized: 'Você não tem autorização para efetuar essa ação',
+      });
+    }
 
     await category.delete();
     return response.status(200).send();
   }
 }
-
 
 module.exports = CategoryController;
